actions/src/py_pydub/requirements.txt
pydub
python-swiftclient
